ensure float division in accuracy calculation
(*(P@#@# pytorch

+++ kraken/lib/train.py
@@ -265,7 +265,7 @@ def recognition_evaluator_fn(model, val_loader, device):
     rec = models.TorchSeqRecognizer(model, device=device)
     chars, error = compute_error(rec, val_loader)
     model.train()
-    accuracy = ((chars-error)/chars).item()
+    accuracy = ((chars-error).float()/chars).item()
     return {'val_metric': accuracy, 'accuracy': accuracy, 'chars': chars, 'error': error}
